conditional compilation works with cost estimator, docker compose up also works
