+++ docker-compose-eigenda.yml
@@ -10,7 +10,7 @@ services:
     ports:
       - "5432:5432"
 
-  op-succinct-celestia:
+  op-succinct-eigenda:
     build:
       context: .
       dockerfile: ./validity/Dockerfile.eigenda
