+++ scripts/utils/Cargo.toml
@@ -64,3 +64,4 @@ op-succinct-build-utils.workspace = true
 [features]
 default = []
 celestia = ["op-succinct-host-utils/celestia", "op-succinct-client-utils/celestia"]
+eigenda = ["op-succinct-host-utils/eigenda", "op-succinct-client-utils/eigenda"]
