+++ utils/client/src/eigenda_client.rs
@@ -1,3 +1,4 @@
+// eigenda_client.rs
 use alloy_consensus::BlockBody;
 use alloy_primitives::{Sealed, B256};
 use alloy_rlp::Decodable;
