+++ utils/client/src/lib.rs
@@ -12,6 +12,7 @@ extern crate alloc;
 
 pub mod client;
 
-pub mod eigenda_client;
-
 pub mod witness;
+
+#[cfg(feature = "eigenda")]
+pub mod eigenda_client;
