+++ validity/Dockerfile.eigenda
@@ -22,6 +22,10 @@ RUN curl -L https://sp1.succinct.xyz | bash && \
 # Copy only what's needed for the build
 COPY . .
 
+# Copy SRS from localhost
+# ToDo ideally from a docker container
+COPY resources /build/resources
+
 # Build the server
 RUN --mount=type=ssh \
     --mount=type=cache,target=/root/.cargo/registry \
@@ -53,6 +57,7 @@ RUN curl -L https://sp1.succinct.xyz | bash && \
 
 # Copy only the built binaries from builder
 COPY --from=builder /build/eigenda-validity-proposer /usr/local/bin/eigenda-validity-proposer
+COPY --from=builder /build/resources  /app/resources
 
 # Set jemalloc flags to aggressively release memory to avoid memory fragmentation.
 ENV JEMALLOC_SYS_WITH_MALLOC_CONF="background_thread:true,narenas:1,tcache:false,dirty_decay_ms:0,muzzy_decay_ms:0,abort_conf:true"
